refactor(tests): split the tests into separate files (#119)
